telegram-database/requirements.txt
telethon==1.32.1
requests==2.31.0
beautifulsoup4==4.12.2
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
python-dotenv==1.0.0
schedule==1.2.0
aiohttp==3.9.1
openai==1.3.7
tavily-python==0.3.1
exa-py==1.8.9
lookup-sdk==0.1.0 